pyproject.toml
[project]
description = "Airflow Sandbox"
readme = "README.md"
requires-python = ">=3.12"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool]

[tool.poetry]
package-mode = false

[tool.poetry.dependencies]
apache-airflow = ">=2.10.2"
python = ">=3.12,<3.13"

[tool.ruff]
src = ["src"]

[tool.ruff.lint]
ignore = [
  "C901", # too complex
]
select = [
  "E", # pycodestyle errors
  "W", # pycodestyle warnings
  "F", # pyflakes
  "I", # isort
  "C", # flake8-comprehensions
  "B", # flake8-bugbear
]

[tool.mypy]
python_version = "3.12"
check_untyped_defs = true
ignore_missing_imports = true
warn_return_any = true
